+++ hutool-core/src/main/java/cn/hutool/core/util/ReferenceUtil.java
@@ -11,8 +11,8 @@ import java.lang.ref.WeakReference;
  * 主要封装包括：
  * <pre>
  * 1. {@link SoftReference} 软引用，在GC报告内存不足时会被GC回收
- * 2. {@link WeakReference}
- * 3. {@link PhantomReference}
+ * 2. {@link WeakReference} 弱引用，在GC时发现弱引用会回收其对象
+ * 3. {@link PhantomReference} 虚引用，在GC时发现虚引用对象，会将{@link PhantomReference}插入{@link ReferenceQueue}。 此时对象未被真正回收，要等到{@link ReferenceQueue}被真正处理后才会被回收。
  * </pre>
  * 
  * @author looly