+++ hutool-db/src/test/java/cn/hutool/db/test/ConcurentTest.java
@@ -31,7 +31,7 @@ public class ConcurentTest {
 	@Before
 	public void init() {
 		LogFactory.setCurrentLogFactory(new ConsoleLogFactory());
-		DbUtil.setShowSqlGlobal(true, false);
+		DbUtil.setShowSqlGlobal(true, false, false);
 		runner = SqlRunner.create();
 	}