+++ hutool-log/src/main/java/cn/hutool/log/dialect/jdk/JdkLogFactory.java
@@ -4,6 +4,7 @@ import java.io.InputStream;
 import java.util.logging.LogManager;
 
 import cn.hutool.core.io.IoUtil;
+import cn.hutool.core.io.resource.ResourceUtil;
 import cn.hutool.core.lang.Console;
 import cn.hutool.log.Log;
 import cn.hutool.log.LogFactory;
@@ -36,7 +37,7 @@ public class JdkLogFactory extends LogFactory{
 	 */
 	private void readConfig() {
 		//避免循环引用，Log初始化的时候不使用相关工具类
-		InputStream in = Thread.currentThread().getContextClassLoader().getResourceAsStream("logging.properties");
+		InputStream in = ResourceUtil.getStreamSafe("logging.properties");
 		if(null == in){
 			System.err.println("[WARN] Can not find [logging.properties], use [%JRE_HOME%/lib/logging.properties] as default!");
 			return;